+++ src/cli/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+    env: {
+        node: true,
+    },
+    plugins: ['node'],
+    extends: ['plugin:node/recommended'],
+};